Centralize types for oldschoolgg web build
