+++ packages/oldschoolgg/src/lib/api.ts
@@ -1,12 +1,16 @@
-import type { IEconomyTransactionsQuery } from '@oldschoolgg/schemas';
+import type {
+	AuthenticatedUser,
+	FullMinionData,
+	IEconomyTransactionsQuery,
+	ServiceStatus,
+	SUserIdentity
+} from '@oldschoolgg/schemas';
 import wretch from 'wretch';
 import { retry } from 'wretch/middlewares';
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
 
 import type { EconomyTransactionsResponse } from '@/components/pages/Account/Staff/types.js';
-import type { AuthenticatedUser, FullMinionData, SUserIdentity } from '../../../robochimp/src/http/api-types.js';
-import type { ServiceStatus } from '../../../robochimp/src/structures/ServiceManager.js';
 
 export type SimpleMinionInfo = {
 	is_ironman: boolean;
