+++ packages/robochimp/package.json
@@ -20,6 +20,7 @@
 		"@hono/node-server": "^1.19.9",
 		"@hono/oauth-providers": "^0.8.5",
 		"@octokit/graphql": "^8.1.1",
+		"@oldschoolgg/schemas": "workspace:*",
 		"@oldschoolgg/toolkit": "workspace:*",
 		"@oldschoolgg/util": "workspace:*",
 		"@prisma/adapter-pg": "^6.19.0",
