+++ packages/robochimp/src/structures/ServiceManager.ts
@@ -1,27 +1,10 @@
-import { AServiceTypes, type IServiceType } from '@oldschoolgg/schemas';
+import { AServiceTypes, type IServiceType, type ServiceStatus } from '@oldschoolgg/schemas';
 import { exec } from 'child_process';
 import { promisify } from 'util';
 
 const execAsync = promisify(exec);
 
 type ServiceAction = 'start' | 'stop' | 'restart' | 'status';
-
-export interface ServiceStatus {
-	service: IServiceType;
-	active: boolean;
-	state: string;
-	sub_state: string;
-	uptime: number | null;
-	memory: {
-		current: number;
-		peak: number;
-	} | null;
-	cpu: {
-		usage: number;
-	} | null;
-	pid: number | null;
-	restart_count: number;
-}
 
 class ServiceManager {
 	private async exec(service: IServiceType, action: ServiceAction): Promise<string> {
